感情分析の結果をクライアントで受け取れるように

+++ deathma-tv-server/src/main/kotlin/tech/notchman/model/Emotions.kt
@@ -4,12 +4,12 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 class Emotions(
-    val joy: Double,
-    val sadness: Double,
-    val anticipation: Double,
-    val surprise: Double,
-    val anger: Double,
-    val fear: Double,
-    val disgust: Double,
-    val trust: Double,
+    val joy: Float,
+    val sadness: Float,
+    val anticipation: Float,
+    val surprise: Float,
+    val anger: Float,
+    val fear: Float,
+    val disgust: Float,
+    val trust: Float,
 )
